company_scraper_copy/requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0 
langchain>=0.1.17
langchain-community>=0.0.5
langchain-tavily>=0.0.5